chore(audit): manual test script + tolerate fenced JSON output
Replaces the in-tree CLI entrypoint with a manual test in test/manual
that calls AuditRunner.run directly, prints the structured audit, and
exits. Strips Markdown ```json fences before JSON.parse so audits no
longer fail when the model wraps output in a code block.
